backend/requirements.txt
# eVoiceClaw Desktop v3 — 精确版本锁定
#
# 此文件由 pip freeze 生成并手工整理，仅包含直接依赖。
# 范围约束和升级说明见 requirements.in。
#
# 平台兼容性说明:
#   - torch==2.2.2: macOS Intel (远程 MacBook Pro) 可用的最高版本
#   - transformers<4.47: 4.47+ 强制要求 torch>=2.6，macOS Intel 无预编译包
#   - numpy<2.0: torch 2.2.x 不兼容 numpy 2.x

# Web 框架
fastapi==0.135.1
uvicorn[standard]==0.41.0

# LLM 适配
litellm==1.82.0

# 向量数据库
lancedb==0.25.3
pyarrow==23.0.1

# Embedding (本地备选，按需启用)
# sentence-transformers==5.2.3

# 配置
pyyaml==6.0.3

# 认证
pyjwt==2.11.0

# HTTP 客户端
httpx==0.28.1

# SSE
sse-starlette==3.3.2

# 隐私管道 — AC 自动机多模式匹配（需 C 编译器，可选）
pyahocorasick>=2.1.0  # 未 pin: 当前环境未安装，需 C 编译器构建

# 隐私管道 — NER 兜底
transformers==4.46.3
torch==2.2.2
numpy==1.26.4

# 工具
python-multipart==0.0.22

# PDF 解析
pdfplumber==0.11.9

# OCR（可选，加载失败时自动降级跳过）
rapidocr-onnxruntime>=1.3.0  # 未 pin: 当前环境未安装，按需安装

# 编译（开发/CI 使用，运行时不需要）
# cython>=3.0
